(CHORES) Split static_specs into submodule
