+++ circuits/src/parsing/mod.rs
@@ -25,4 +25,4 @@ mod table;
 pub use base64_chip::*;
 pub use data_types::{DateFormat, Separator};
 pub use parser_gadget::*;
-pub use scanner::{regex, spec_library, StdLibParser};
+pub use scanner::{regex, static_specs::spec_library, StdLibParser};
