--- circuits/src/parsing/scanner/static_specs.rs
@@ -1,770 +1,0 @@
-//! A library of parsers, represented as a Hash Map from documented fixed
-//! tokens (`StdLibParser`) to deterministic minimal automata (`Automata`).
-//! Since these parsers can be relatively costly to generate, the automata are
-//! serialized. The tests check in particular the consistency between the
-//! serialised data and a freshly computed one.
-//!
-//! Whenever a new parser has to be added to the parser library, the following
-//! steps have to be followed:
-//! 1. Add a corresponding entry (and documentation) in the `StdLibParser` type.
-//! 2. Define a function `spec_*` returning the `Regex` defining the parser.
-//! 3. Create an empty file whose name and location matches the result of
-//!    `parser.serialization_file()`, where `parser` is the entry defined at
-//!    step 1.
-//! 4. Add an entry `(parser, spec, serialization)` in the function
-//!    `spec_library_data`, following the format of the other entries. The
-//!    `serialization` component is, in particular, an `include_bytes!` of the
-//!    file created at step 3.
-//! 5. Run the tests of this file from the root of the repository. This will
-//!    bootstrap the serialisation file created at step 1.
-//! 6. If the serialisation data needs to be updated, truncate the content of
-//!    the corresponding serialisation file, and re-run step 5.
-//!
-//! ```shell
-//!    cargo test --lib -p midnight-circuits --release -- --nocapture regex_test automaton_test static_specs_test
-//! ```
-
-#[cfg(test)]
-use std::{fs, path::Path};
-
-use rustc_hash::FxHashMap;
-
-use super::{
-    automaton::Automaton,
-    regex::{Regex, RegexInstructions},
-    serialization::Serialize,
-};
-
-/// Folder where the serialized automata for the standard library will be
-/// stored.
-#[cfg(test)]
-const AUTOMATON_CACHE: &str = "src/parsing/scanner/automaton_cache";
-
-/// Explicit names (and documentation) for indexing the various parsing
-/// specifications hard-coded in the standard library.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub enum StdLibParser {
-    /// # Description and sources
-    ///
-    /// A JWT (Json Web Token) credential payload format, in compliance with the
-    /// [RFC 7519](https://datatracker.ietf.org/doc/html/rfc7519) of the IETF. It uses
-    /// [this data model in the VC field](https://www.w3.org/TR/vc-data-model-2.0/).
-    ///
-    /// Notes:
-    ///   - Compliance with RFC 7519: the optional fields "iat" and "jti" are
-    ///     not checked by this credential. The fields "iss", "sub", "nbf" and
-    ///     "exp" are required by this credential.
-    ///   - The parser only accepts fields *in the same order as below*.
-    ///
-    /// ```text
-    /// {
-    ///     iss: string,
-    ///     sub: string,
-    ///     nbf: number,
-    ///     exp: number,
-    ///     vc: {
-    ///       credentialSchema?: {
-    ///         id: string,
-    ///         type: string
-    ///       }[],
-    ///       credentialSubject: {
-    ///         nationalId: string, # output "1"
-    ///         familyName: string, # output "2"
-    ///         givenName: string,  # output "3"
-    ///         publicKeyJwk: {
-    ///           kty: string,
-    ///           crv: string,
-    ///           x: string, # output "5"
-    ///           y: string  # output "6"
-    ///         }
-    ///         id: string,
-    ///         birthDate: string,  # output "4"
-    ///       },
-    ///       type: string[],
-    ///       @context: string[],
-    ///       issuer: string | { id: string, type?: string },
-    ///       credentialStatus: {
-    ///         statusPurpose: string,
-    ///         statusListIndex: number,
-    ///         id: string,
-    ///         type: string,
-    ///         statusListCredential: string
-    ///       }
-    ///     }
-    /// }
-    /// ```
-    ///
-    /// In particular, the `string` token above refers to JSON strings. This
-    /// parser notably enforces the low-level requirements of JSON strings,
-    /// such as being UTF-8 encoded or correctly escaped, as required in
-    /// [RFC 8259](https://datatracker.ietf.org/doc/html/rfc8259) §7.
-    ///
-    /// # Output Behaviour:
-    ///
-    /// As Specified in the above grammar, the following field contents
-    /// (excluding the double quotes) are output as follows:
-    ///   - `"nationalId"` -> 1
-    ///   - `"familyName"` -> 2
-    ///   - `"givenName"` -> 3
-    ///   - `"birthDate"` -> 4
-    ///   - `"x"` -> 5
-    ///   - `"y"` -> 6
-    Jwt,
-
-    /// # Description and sources
-    ///
-    /// Format of the Data Group 1 (DG1) of biometric passports, as specified in
-    /// ICAO Doc 9303 for TD3-type documents (machine-readable passports).
-    ///
-    /// The DG1 contains the Machine Readable Zone (MRZ) printed on the
-    /// passport's main page and stored verbatim on the embedded chip. It
-    /// encodes key identity attributes such as the (possibly truncated)
-    /// holder's name, date of birth, nationality, and passport number,
-    /// using a fixed-length ASCII format.
-    ///
-    /// A typical TD3 MRZ looks as follows on passports:
-    ///
-    /// ```text
-    /// PPFRADUPONT<<JEAN<MICHEL<<<<<<<<<<<<<<<<<<<<
-    /// 12AB345678FRA7408122M3101012<<<<<<<<<<<<<<04
-    /// ```
-    ///
-    /// Note that despite being presented as two lines of 44 bytes each, the DG1
-    /// is read and parsed *as a single 88-byte line*.
-    ///
-    /// These 88 bytes, along with the other Data Groups, can be retrieved by
-    /// reading the passport's NFC chip. Integrity and authenticity are ensured
-    /// via the Security Object Document (SOD), which contains the signed
-    /// sequence of hashes of the Data Groups. These signatures can be
-    /// verified using public keys distributed through the ICAO Public Key
-    /// Directory (PKD).
-    ///
-    /// Sources:
-    /// - [ICAO Doc 9303](https://www.icao.int/publications/doc-series/doc-9303)
-    /// - [PKD](https://www.icao.int/icao-pkd)
-    ///
-    /// # Output behaviour
-    ///
-    /// In the MRZ format, only uppercase letters, digits, and `<` (representing
-    /// special characters such as spaces or dashes, as well as padding and
-    /// separators), are used. The following fields are then output as follows;
-    /// note that checksum fields are not mentioned, as they get no output and
-    /// are not verified by the parser.
-    ///  - Passport type (2 bytes; uppercase) -> 1
-    ///  - Issuing country code (3 bytes; uppercase) -> 2
-    ///    + **Note**: this code is ISO 3166-1 alpha-3 compliant.
-    ///  - name field (up to 39 bytes; uppercase and spaces) -> 3 (surname) and
-    ///    4 (given names, if any)
-    ///    + **Note 1**: mononyms, i.e., people having no given names, are
-    ///      ICAO9303 TD3 compliant. In this case, no byte will be output `4`.
-    ///    + **Note 2**: if the credential holder has at least one given name,
-    ///      the credential must include a `<<` separator between the surname
-    ///      and the given names. Names may be truncated if needed to make the
-    ///      separator fit. The whole field is also padded with `<` bytes if no
-    ///      truncation occurred.
-    ///    + **Note 3**: The `<` characters (separator or padding) are not
-    ///      output by this parser. E.g., in `DUPONT<<JEAN<MICHEL`, the 3 `<`
-    ///      get no output. This is because of the impossibility, for a finite
-    ///      automaton to, e.g., decide the third `<` is a padding or, like
-    ///      here, a space before an additional given name.
-    ///  - Passport number (9 bytes; uppercase and digits) -> 5
-    ///  - Nationality (3 bytes; uppercase) -> 6
-    ///  - Date of birth (6 bytes; YYMMDD) -> 7
-    ///  - Sex (1 byte; `M` Male, `F` Female, or `<` Other) -> 8
-    ///  - Date of expiry (6 bytes; YYMMDD) -> 9
-    ///  - Optional data (14 bytes; uppercase, digits, and `<`) -> 10
-    Icao9309Td3Dg1,
-}
-
-#[cfg(test)]
-impl StdLibParser {
-    /// Returns the file name where the parser index is serialised. The path
-    /// starts from the root of the repository.
-    pub(super) fn serialization_file(&self) -> String {
-        format!("{}/{:?}", AUTOMATON_CACHE, self)
-    }
-}
-
-/// The raw entry of the parsing library. Contains the different parser names
-/// (`StdLibParser`), the functions used to generate the corresponding `Regex`
-/// (functions typically named `spec_*`), and the serialization bytes.
-type LibraryData = &'static [(StdLibParser, &'static dyn Fn() -> Regex, &'static [u8])];
-
-/// A library of parsing automata, computed or deserialised.
-type ParsingLibrary = FxHashMap<StdLibParser, Automaton>;
-
-/// The basic, non computed data of the parsing library. When serialization is
-/// enabled, the automata will be deserialized in `spec_library` from the third
-/// components. When serialization is disabled, the automaton will be computed
-/// using the second argument.
-fn spec_library_data() -> LibraryData {
-    &[
-        (
-            StdLibParser::Jwt,
-            &spec_jwt,
-            include_bytes!("automaton_cache/Jwt"),
-        ),
-        (
-            StdLibParser::Icao9309Td3Dg1,
-            &spec_icao9303_td3_dg1,
-            include_bytes!("automaton_cache/Icao9309Td3Dg1"),
-        ),
-    ]
-}
-
-/// All automata that can be used as a parsing basis in the standard library.
-/// Exclusively uses serialised data.
-pub fn spec_library() -> ParsingLibrary {
-    spec_library_data()
-            .iter()
-            .map(|(name, _, serialization)| {
-                assert!(
-                    !serialization.is_empty(),
-                    "Empty serialisation data for {:?}. The bootstrapping of the serialisation process has not been conducted. (see documentation of `midnight_circuits::parsing::scanner::static_specs`)",
-                    *name
-                );
-                (*name, Automaton::deserialize_unwrap(serialization))
-            })
-            .collect::<FxHashMap<_, _>>()
-}
-
-/// Regex formalising the spec of `StdLIbParser::Jwt`.
-fn spec_jwt() -> Regex {
-    // Content of a basic field (RFC 8259 JSON string), possibly with output if
-    // `output` is not 0.
-    let string = |output: usize| -> Regex {
-        Regex::json_string().replace_outputs(&|m| if m == 1 { Some(output) } else { None })
-    };
-    // A json field, with possible white spaces.
-    let field = |name: &str, content: Regex| -> Regex {
-        Regex::spaced_cat([format!("\"{name}\"").into(), ":".into(), content])
-    };
-    let string_field = |name: &str, output: usize| -> Regex { field(name, string(output)) };
-    let int_field = |name: &str| -> Regex { field(name, Regex::digit().non_empty_list()) };
-
-    // A collection of regular expressions, delimited by `opening` and `closing`,
-    // and separated by commas. Arbitrary white spaces are allowed between the
-    // delimiters and the various items.
-    let collec = |opening: &str, items: Vec<Regex>, closing: &str| -> Regex {
-        Regex::spaced_separated_cat(items, ",".into())
-            .spaced_delimited(opening.into(), closing.into())
-    };
-
-    // A JSON list of strings.
-    let string_list = string(0)
-        .spaced_separated_list(",".into())
-        .spaced_delimited("[".into(), "]".into());
-
-    // Fields of the credential format.
-    let credential_schema = field(
-        "credentialSchema",
-        collec(
-            "{",
-            vec![string_field("id", 0), string_field("type", 0)],
-            "}",
-        )
-        .spaced_separated_list(",".into())
-        .spaced_delimited("[".into(), "]".into()),
-    );
-    let public_key_jwk = field(
-        "publicKeyJwk",
-        collec(
-            "{",
-            vec![
-                string_field("kty", 0),
-                string_field("crv", 0),
-                string_field("x", 5), // Output 5.
-                string_field("y", 6), // Output 6.
-            ],
-            "}",
-        ),
-    );
-    let credential_subject = field(
-        "credentialSubject",
-        collec(
-            "{",
-            vec![
-                string_field("nationalId", 1), // Output 1.
-                string_field("familyName", 2), // Output 2.
-                string_field("givenName", 3),  // Output 3.
-                public_key_jwk,
-                string_field("id", 0),
-                string_field("birthDate", 4), // Output 4.
-            ],
-            "}",
-        ),
-    );
-    let issuer = field(
-        "issuer",
-        Regex::union([
-            string(0),
-            collec(
-                "{",
-                vec![string_field("id", 0), string_field("type", 0)],
-                "}",
-            ),
-            collec("{", vec![string_field("id", 0)], "}"),
-        ]),
-    );
-    let credential_status = field(
-        "credentialStatus",
-        collec(
-            "{",
-            vec![
-                string_field("statusPurpose", 0),
-                int_field("statusListIndex"),
-                string_field("id", 0),
-                string_field("type", 0),
-                string_field("statusListCredential", 0),
-            ],
-            "}",
-        ),
-    );
-
-    collec(
-        "{",
-        vec![
-            string_field("iss", 0),
-            string_field("sub", 0),
-            int_field("nbf"),
-            int_field("exp"),
-            field(
-                "vc",
-                collec(
-                    "{",
-                    vec![
-                        credential_subject.clone().or(Regex::spaced_cat([
-                            credential_schema,
-                            ",".into(),
-                            credential_subject,
-                        ])),
-                        field("type", string_list.clone()),
-                        field("@context", string_list),
-                        issuer,
-                        credential_status,
-                    ],
-                    "}",
-                ),
-            ),
-        ],
-        "}",
-    )
-}
-
-/// Regex formalising the spec of `StdLIbParser::ICAO9303DataGroup1`.
-fn spec_icao9303_td3_dg1() -> Regex {
-    // The list of all tolerated passport types. Consulted at this document, Section
-    // 4.4, on Jan. 14, 2026:
-    // https://www.icao.int/sites/default/files/publications/DocSeries/9303_p4_cons_en.pdf
-    // Output 1.
-    let passport_type = Regex::union([
-        "P<".into(), // Legacy denomination of `PP`, before Jan. 2026.
-        "PP".into(), // National/Ordinary passport.
-        "PE".into(), // Emergency passport.
-        "PD".into(), // Diplomatic passport.
-        "PO".into(), // Official/Service passport.
-        "PR".into(), // Refugee passport.
-        "PT".into(), // Alien passport.
-        "PS".into(), // Stateless passport.
-        "PL".into(), // Laissez-passez passport.
-        "PM".into(), // Military passport.
-        "PU".into(), /* Emergency travel document. See: https://www.icao.int/sites/default/files/publications/DocSeries/9303_p8_cons_en.pdf */
-    ]).output(&|_| Some(1));
-    // A non-empty sequence of uppercase letters, with output `output`.
-    let name_block = |output: usize| -> Regex {
-        Regex::uppercase_letter().non_empty_list().output(&|_| Some(output))
-    };
-    // One uppercase letter or a digit, with output `output`.
-    let alphanum = |output: usize| -> Regex {
-        Regex::byte_from((b'A'..=b'Z').chain(b'0'..=b'9')).output(&|_| Some(output))
-    };
-    // A date with the given output, in YYMMDD format.
-    let date = |output: usize| -> Regex { Regex::digit().output(&|_| Some(output)).repeat(6) };
-    // Any passport character.
-    let any = Regex::byte_from((b'A'..=b'Z').chain(b'0'..=b'9').chain(std::iter::once(b'<')));
-
-    // Example to illustrate the code below:
-    // P<FRADUPONT<<JEAN<MICHEL<<<<<<<<<<<<<<<<<<<<
-    // 12AB345678FRA7408122M3101012<<<<<<<<<<<<<<04
-
-    // Mandatory part of the first line of the DG1 (passport type, issuer, surname).
-    // The separators `<` get no output.
-    let line1_prefix = Regex::cat([
-        passport_type,
-        Regex::uppercase_letter().output(&|_| Some(2)).repeat(3),
-        name_block(3).separated_non_empty_list("<".into()),
-    ]);
-    // Given names in the first line of the DG1, prefixed with the separator. The
-    // separators `<` get no output.
-    let given_names = Regex::cat([
-        "<<".into(),
-        name_block(4).separated_non_empty_list("<".into()),
-    ]);
-    // The part of the first line following `line1_prefix`. Considers cases where
-    // given names are present or not.
-    let line1_suffix = given_names.optional().terminated(Regex::word("<").list());
-    // The full first line, with the length constraint.
-    let line1 = line1_prefix.terminated(line1_suffix).and(any.clone().repeat(44));
-
-    // The second line of the DG1. All fields are length constrained.
-    let line2 = Regex::cat([
-        alphanum(5).repeat(9),
-        Regex::digit(), // No-output checksum.
-        Regex::uppercase_letter().output(&|_| Some(6)).repeat(3),
-        date(7),
-        Regex::digit(), // No-output checksum.
-        Regex::byte_from([b'<', b'M', b'F']).output(&|_| Some(8)),
-        date(9),
-        Regex::digit(), // No-output checksum.
-        any.repeat(14).output(&|_| Some(10)),
-        Regex::digit().repeat(2), // No-output checksum.
-    ]);
-
-    // Concatenating the two lines, without a newline character.
-    line1.terminated(line2)
-}
-
-#[cfg(test)]
-/// Re-serialises the data in `checks`, and:
-///
-/// 1. If some non-empty serialisation file exists and
-///    `AUTOMATON_BREAKING_CHANGE` is set to false, panics if the serialized
-///    data is inconsistent.
-/// 2. If some empty serialisation file exists, writes the serialised data
-///    inside.
-/// 3. If the expected serialisation files do not exist, panics.
-///
-/// Will also panic and request a re-compilation if the serialisation data is
-/// updated at any point.
-///
-/// Note: This function is only available in tests.
-fn check_serialization(checks: &ParsingLibrary) {
-    // Tracks whether a recompilation is needed so that the serialisation data is in
-    // sync.
-    let mut recompile = false;
-    for (parser, automaton) in checks {
-        let file_name = parser.serialization_file();
-        assert!(
-                Path::new(&file_name).exists(),
-                "serialisation file {file_name} does not exist! Follow the documentation of `midnight_circuits::parsing::scanner::static_specs` for instructions on how to add a new parser to the standard library."
-            );
-        let previous_data = fs::read(file_name.clone()).unwrap();
-        let mut current_data = Vec::new();
-        automaton.serialize(&mut current_data);
-        if previous_data.is_empty() {
-            println!("-> bootstrapping the serialisation of {:?}. Recompilation will be necessary so that the executable contains the correct serialised data.", parser);
-            recompile = true;
-            fs::write(file_name, &current_data).unwrap();
-        } else {
-            assert!(
-                current_data == previous_data,
-                "The serialisation data of the parsing library (parser name: {:?}) is not up to date. If this is intentional, clear the content of {}, and run the test again to replace its content.",
-                parser, parser.serialization_file()
-            );
-            println!("-> serialisation data of {:?} is up to date.", parser)
-        }
-        println!(">> Serialisation checks completed.\n======");
-        assert!(
-            !recompile,
-            "The executable has to be re-compiled so that the serialisation data is up-to-date."
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::time::Instant;
-
-    use rustc_hash::{FxBuildHasher, FxHashMap};
-
-    use super::{
-        super::automaton::Automaton, check_serialization, spec_library, spec_library_data,
-        StdLibParser,
-    };
-
-    /// Sets up the serialised library (bootstraps it if empty serialisation
-    /// data is found), and performs consistency checks or updates accordingly
-    /// to the value of `AUTOMATON_BREAKING_CHANGE`.
-    fn configure_serialisation() {
-        let lib_data = spec_library_data();
-
-        println!("======\nRecomputing the parsing library automata...");
-        let mut lib = FxHashMap::with_capacity_and_hasher(lib_data.len(), FxBuildHasher);
-        let start = Instant::now();
-        for (name, spec, _) in lib_data {
-            let start_local = Instant::now();
-            let automaton = spec().to_automaton();
-            println!(
-                "-> Generated {:?} automaton in {:?}",
-                *name,
-                start_local.elapsed()
-            );
-            lib.insert(*name, automaton);
-        }
-        println!(
-            ">> Full parsing library re-computed in {:?}!\n======\n>> Now checking the consistency of serialised data.",
-            start.elapsed()
-        );
-        check_serialization(&lib)
-    }
-
-    /// Tests whether a given regular expression accepts or rejects two sets of
-    /// corresponding strings. For accepted strings, checks the list of outputs
-    /// for each output.
-    fn specs_one_test(
-        spec_library: &FxHashMap<StdLibParser, Automaton>,
-        spec: StdLibParser,
-        accepted: &[(&str, &[(usize, &str)])],
-        rejected: &[&str],
-    ) {
-        let automaton = spec_library.get(&spec).unwrap();
-        println!("\n\n** TEST of the spec {:?}", spec);
-        accepted.iter().enumerate().for_each(|(index,&(s,expected_outputs))| {
-            println!("\n -> accepting test nb. {index}");
-            let s_bytes = s.as_bytes();
-
-            let (v,output_automaton,interrupted) = automaton.run(s_bytes);
-            let counter = v.len() - 1;
-            assert!(!interrupted,
-                "input was unexpectedly rejected after being stuck after {} transitions, reading character '{}' (byte {}). Partial input read before the interruption:\n\n{}\n\n(i.e., bytes [{:?}])",
-                counter,
-                s_bytes[counter] as char,
-                s_bytes[counter],
-                &s[..counter],
-                &s_bytes[..counter],
-            );
-
-            // Gathering outputs.
-            let mut outputs = FxHashMap::with_capacity_and_hasher(2, FxBuildHasher);
-            for (&o,&i) in output_automaton.iter().zip(s_bytes) {
-                if o != 0 {
-                    outputs.entry(o).or_insert(vec![]).push(i);
-                }
-            }
-
-            let state = v[counter];
-            assert!(
-                automaton.final_states.contains(&state),
-                "input was unexpectedly rejected (automaton run ended up in the non-final state {} after {} transitions)", state, counter
-            );
-            if let Some(n) = outputs.iter().find(|&(i,_)|
-                expected_outputs.iter().all(|(j,_)| i != j)
-            ){
-                panic!(
-                    "[test of spec {:?}, nb. {index}]: the input {s} is accepted as expected, but it has an unexpected output {}, which is unexpected\nThe automaton reached the final state {} in {} transitions.",
-                    spec, n.0, state, counter
-                )
-            }
-            for (i,expected_output) in expected_outputs {
-                let expected_output_bytes = expected_output.as_bytes();
-                match outputs.get(i) {
-                    None => panic!(
-                        "[test of spec {:?}, nb. {index}]: the input {s} is accepted as expected, but it has no output {i}, which is unexpected\nThe automaton reached the final state {} in {} transitions.",
-                        spec, state, counter
-                    ),
-                    Some(output_bytes) => {
-                        assert!(output_bytes == expected_output_bytes,
-                            "[test of spec {:?}, nb. {index}]: the input {s} is accepted as expected, but output {i} is\n  \"{}\"\ninstead of\n  \"{}\"\nwhich is unexpected. The automaton reached the final state {} in {} transitions.",
-                            spec,
-                            String::from_utf8_lossy(output_bytes),
-                            String::from_utf8_lossy(expected_output_bytes),
-                            state,
-                            counter
-                        );
-
-                    }
-                }
-            }
-            println!("... which is accepted as expected with the correct outputs (the automaton reached the final state {} in {} transitions). The outputs are:", state, counter);
-            for (i,o) in expected_outputs {
-                println!("  - {i}: {o}")
-            }
-        });
-        rejected.iter().enumerate().for_each(|(index,s)| {
-            println!("\n -> rejecting test nb. {index}");
-            let s_bytes = s.as_bytes();
-            let (v,output,interrupted) = automaton.run(s_bytes);
-            let counter = v.len() - 1;
-            if interrupted {
-                println!(
-                    "... which is rejected as expected. The automaton run was stuck after {} transitions, reading character '{}' (byte {}). Partial input read before the interruption:\n\n{}\n",
-                    counter,
-                    s_bytes[counter] as char,
-                    s_bytes[counter],
-                    &s[..counter],
-                )
-            } else {
-                let state = v[counter];
-                if automaton.final_states.contains(&state) {
-                    // Gathering outputs.
-                    let mut outputs = FxHashMap::with_capacity_and_hasher(2, FxBuildHasher);
-                    for (&o,&i) in output.iter().zip(s_bytes) {
-                        if o != 0 {
-                            outputs.entry(o).or_insert(vec![]).push(i);
-                        }
-                    }
-                    let mut outputs_str = String::new();
-                    for (i,o) in outputs {
-                        outputs_str.push_str(&format!("  - {i}: {}\n", String::from_utf8_lossy(&o)))
-                    }
-                    panic!(
-                        "input was unexpectedly accepted (reached final state {} after {} transitions). The outputs were:\n{}",
-                        state, counter, outputs_str
-                    )
-                }
-                println!("... which is rejected as expected (the automaton run ended up in the non-final state {} after {} transitions).", state, counter)
-            }
-        });
-    }
-
-    #[test]
-    fn specs_test() {
-        // Enforces that serialisation data is consistent and up to date.
-        configure_serialisation();
-
-        // Performs the tests using the serialised data.
-        println!(">> Now configuring the spec library for tests... (using the serialised data)");
-        let start = Instant::now();
-        let spec_library = spec_library();
-        println!(
-            ">> Configuration completed in {:?}. Automaton breakdown:",
-            start.elapsed()
-        );
-        let mut total = 0;
-        for (name, automaton) in &spec_library {
-            println!(
-                "  - {:?}: {} states, {} transitions",
-                name,
-                automaton.nb_states,
-                automaton.transitions.values().map(|m| m.len()).sum::<usize>()
-            );
-            total += automaton.transitions.values().map(|m| m.len()).sum::<usize>()
-                + automaton.final_states.len()
-        }
-        println!(
-            ">> Total nb of lookup rows in the chip: {} ≤ 2^{}",
-            total,
-            total.next_power_of_two().trailing_zeros()
-        );
-
-        // Tests the `Jwt` spec correctness.
-        const FULL_INPUT_JWT: &str = include_str!("specs_examples/jwt/full.txt");
-        const MINIMAL_JWT: &str = include_str!("specs_examples/jwt/minimal.txt");
-        let accepted0: Vec<(&str, &[(usize, &str)])> = vec![
-            (
-                FULL_INPUT_JWT,
-                &[
-                    (1, "12345"),
-                    (2, "Wonderland"),
-                    (3, "Alice"),
-                    (4, "2000-11-13"),
-                    (5, "S0kj3ydSeF86LU9BpHuVntMFN8SCKcHyci1tXFbRW8M"),
-                    (6, "dux8h-QcIA3aZG9CSPIltDwVvOkf0kfJRJLH7K1KSlQ"),
-                ],
-            ),
-            (
-                MINIMAL_JWT,
-                &[
-                    (1, "id"),
-                    (2, "fn"),
-                    (3, "gn"),
-                    (4, "bd"),
-                    (5, "x"),
-                    (6, "y"),
-                ],
-            ),
-        ];
-        let rejected0: Vec<&str> =
-            vec!["hello world", &FULL_INPUT_JWT[..1000], &MINIMAL_JWT[..600]];
-        specs_one_test(&spec_library, StdLibParser::Jwt, &accepted0, &rejected0);
-
-        // Tests the `ICAO9303DataGroup1` spec correctness.
-        let accepted1_raw = include_str!("specs_examples/icao9303_td3_dg1/valid_credentials.txt")
-            .lines()
-            .collect::<Vec<_>>();
-        let accepted1: Vec<(&str, &[(usize, &str)])> = vec![
-            (
-                accepted1_raw[0],
-                &[
-                    (1, "PP"),
-                    (2, "JPN"),
-                    (3, "OKABE"),
-                    (4, "RINTARO"),
-                    (5, "12AB34567"),
-                    (6, "JPN"),
-                    (7, "911214"),
-                    (8, "M"),
-                    (9, "310101"),
-                    (10, "EL<PSY<CONGROO"),
-                ],
-            ),
-            (
-                accepted1_raw[1],
-                &[
-                    (1, "PE"),
-                    (2, "ESP"),
-                    (3, "DELACRUZ"),
-                    (4, "MARIA"),
-                    (5, "UH87G9901"),
-                    (6, "ESP"),
-                    (7, "911214"),
-                    (8, "F"),
-                    (9, "310101"),
-                    (10, "XXV789<<<<<<<<"),
-                ],
-            ),
-            (
-                accepted1_raw[2],
-                &[
-                    (1, "PD"),
-                    (2, "MDG"),
-                    (3, "ANDRIANAMPOINIMERINATOMPOLOINDRINDRA"),
-                    (4, "R"),
-                    (5, "BDL3820HR"),
-                    (6, "FRA"),
-                    (7, "450101"),
-                    (8, "<"),
-                    (9, "600101"),
-                    (10, "<<<<<<<<<<<<<<"),
-                ],
-            ),
-            (
-                accepted1_raw[3],
-                &[
-                    (1, "PO"),
-                    (2, "FRA"),
-                    (3, "NOOOWAYIGOTATRUNCATEDMONONYMRIGH"),
-                    (5, "AAAAAAAAA"),
-                    (6, "FRA"),
-                    (7, "990101"),
-                    (8, "<"),
-                    (9, "300101"),
-                    (10, "<<<<<<<<<<<<<<"),
-                ],
-            ),
-            (
-                accepted1_raw[4],
-                &[
-                    (1, "PR"),
-                    (2, "USA"),
-                    (3, "ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ"),
-                    (5, "PPPPPPPPP"),
-                    (6, "USA"),
-                    (7, "990101"),
-                    (8, "M"),
-                    (9, "300102"),
-                    (10, "<<<<<<<<<<<<<<"),
-                ],
-            ),
-        ];
-        let rejected1 = include_str!("specs_examples/icao9303_td3_dg1/invalid_credentials.txt")
-            .lines()
-            .collect::<Vec<_>>();
-        specs_one_test(
-            &spec_library,
-            StdLibParser::Icao9309Td3Dg1,
-            &accepted1,
-            &rejected1,
-        );
-    }
-}
