+++ circuits/src/parsing/scanner/static_specs/examples/icao9303_td3_dg1/gen.rs
