+++ circuits/src/parsing/scanner/static_specs/examples/jwt/full.txt
