+++ circuits/src/parsing/scanner/static_specs/examples/jwt/minimal.txt
